requirement.txt
asn1crypto==0.24.0
blinker==1.4
cffi==1.12.3
Click==7.0
colour-printing==0.0.8
cryptography==2.7
ctpbee==0.24
dataclasses==0.6
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
motor==2.0.0
numpy==1.17.0
pandas==0.25.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.6
tornado==6.0.3
Werkzeug==0.15.5
